fix names files
